+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Fixed
+
+- Fix server/channel settings viewer permissions when viewer is owner.
+
 ## [v20.13.0](https://github.com/lexicalunit/spellbot/releases/tag/v20.13.0) - 2026-06-11
 
 ### Added